Refactor the controller codes
