+++ server/commonHelpers.js
