+++ server/controllers/centers/index.js
@@ -1,0 +1,3 @@
+import centerController from './centers';
+
+export default centerController;
