+++ server/controllers/events/index.js
@@ -1,0 +1,3 @@
+import eventController from './events';
+
+export default eventController;
