+++ server/controllers/users/index.js
@@ -1,0 +1,3 @@
+import userController from './users';
+
+export default userController;
